# anti-fraude/src/swagger/anti-fraude.yaml
openapi: 3.0.3
info:
  title: ANTI-FRAUDE API
  description: |-
    API do microsserviço de Anti-fraude
  version: 0.0.1
servers:
  - url: http://localhost:3000
paths:
...
  /api/admin/accounts/{id}:
    get:
      security: 
      - auth: []
      tags:
        - "accounts"
      summary: Find account
      description: Find one specific account
      parameters:
        - name: id
          in: path
          description: ID of account that needs to be fetched
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          description: invalid ID supplied
        "404":
          description: not found
    put:
      security: 
      - auth: []
      tags:
        - "accounts"
      summary: Update account
      description: Update account by entering the whole updated account
      parameters:
        - name: id
          in: path
          description: ID of account that needs to be updated
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Account"
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          description: invalid ID supplied
        "401":
          description: unauthorized
    delete:
      security: 
      - auth: []
      tags:
        - "accounts"
      summary: Delete account
      description: Delete account by entering its ID
      parameters:
        - name: id
          in: path
          description: ID of account that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        "204":
          description: no content
        "401":
          description: unauthorized
        "400":
          description: invalid ID supplied
...
components:
  schemas:
    Account:
      type: object
      properties:
        nome:
          type: string
        email:
          type: string
          format: email
        senha:
          type: string
          format: password
...
  securitySchemes:
    auth:
      type: http
      scheme: bearer
      bearerFormat: JWT